+++ lib/models/filters/filter.dart
@@ -1,0 +1,5 @@
+abstract class Filter{
+
+  Map<String, dynamic> toMap();
+
+}