+++ lib/services/authentication_service.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:http/http.dart';
+import 'package:http/http.dart' as http;
 import 'package:rocket_chat_connector_flutter/models/authentication.dart';
 import 'package:rocket_chat_connector_flutter/models/user.dart';
 import 'package:rocket_chat_connector_flutter/services/http_service.dart';
@@ -12,7 +12,7 @@ class AuthenticationService {
 
   Future<Authentication> login(String user, String password) async {
     Map<String, String> body = {'user': user, 'password': password};
-    Response response = await _httpService.post('/api/v1/login', jsonEncode(body));
+    http.Response response = await _httpService.post('/api/v1/login', jsonEncode(body));
 
     if (response?.statusCode == 200 && response.body?.isNotEmpty == true) {
       return Authentication.fromMap(jsonDecode(response.body));
@@ -21,7 +21,7 @@ class AuthenticationService {
   }
 
   Future<User> me() async {
-    Response response = await _httpService.get('/api/v1/me');
+    http.Response response = await _httpService.get('/api/v1/me');
 
     if (response?.statusCode == 200 && response.body?.isNotEmpty == true) {
       return User.fromMap(jsonDecode(response.body));