+++ lib/services/room_service.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:http/http.dart';
+import 'package:http/http.dart' as http;
 import 'package:rocket_chat_connector_flutter/models/new/room_new.dart';
 import 'package:rocket_chat_connector_flutter/models/response/room_new_response.dart';
 import 'package:rocket_chat_connector_flutter/models/room.dart';
@@ -15,7 +15,7 @@ class RoomService {
   RoomService(this._httpService);
 
   Future<RoomNewResponse> create(RoomNew roomNew) async {
-    Response response = await _httpService.post('/api/v1/im.create', jsonEncode(roomNew.toMap()));
+    http.Response response = await _httpService.post('/api/v1/im.create', jsonEncode(roomNew.toMap()));
 
     if (response?.statusCode == 200 && response.body?.isNotEmpty == true) {
       return RoomNewResponse.fromMap(jsonDecode(response.body));
@@ -24,7 +24,7 @@ class RoomService {
   }
 
   Future<RoomMessages> messages(Room room) async {
-    Response response = await _httpService.get('/api/v1/im.messages?roomId=${room.rid}');
+    http.Response response = await _httpService.get('/api/v1/im.messages?roomId=${room.rid}');
 
     if (response?.statusCode == 200 && response.body?.isNotEmpty == true) {
       return RoomMessages.fromMap(jsonDecode(response.body));
@@ -33,7 +33,7 @@ class RoomService {
   }
 
   Future<RoomCounters> counters(Room room, [User user]) async {
-    Response response;
+    http.Response response;
     if (user != null) {
       response = await _httpService.get('/api/v1/im.counters?roomId=${room.id}&userId=${user.id}');
     } else {